[clutter-event-win32] Fix the argument types to clutter_actor_get_size
clutter_actor_get_size now takes float pointers so it was issuing a
warning.

+++ clutter/win32/clutter-event-win32.c
@@ -359,7 +359,7 @@ message_translate (ClutterBackend *backend,
 	{
 	  WORD new_width = LOWORD (msg->lParam);
 	  WORD new_height = HIWORD (msg->lParam);
-	  guint old_width, old_height;
+	  gfloat old_width, old_height;
 
 	  clutter_actor_get_size (CLUTTER_ACTOR (stage),
 				  &old_width, &old_height);
